requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.2.0
sqlalchemy==2.0.32
sqlalchemy-utils==0.41.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
weasyprint==62.3
python-docx==1.1.2
gunicorn==22.0.0

